splitting out helper tests
